Major fix
I was getting a bit of drift whenever this library was being used in
anything. The problem was ultimately some small timing issues, and also
a delay of one half bit was needed between sending each byte. That
forces the Arduino to read from the middle of a bit, so any fuzzy
transitions wouldn't cause a full bit shift to happen.